qfbn_icml26/requirements.txt
torch>=2.3
torchvision>=0.18
numpy>=1.24
tqdm>=4.66
pyyaml>=6.0
matplotlib>=3.8
